python/requirements.txt
numpy>=1.20.0
Pillow>=9.0.0
matplotlib>=3.5.0
jupyter>=1.0.0
